refactor: create 3 different transform streams
