+++ my_ciphering_cli.js
@@ -1,12 +1,12 @@
-const fs = require('fs');
 const { stdout, stderr, exit } = process;
 const { validateConfig, validateOptions } = require('./helpers/validator');
 const { argsParser } = require('./helpers/parser');
-const CipherTransphorm = require('./stream/cipherTransphorm');
-const CustomWritableStream = require('./stream/customWritableStream');
-const CustomReadableStream = require('./stream/customReadableStream');
-const customStdin = require('./stream/customStdin');
-const { ERROR_MESSAGE } = require('./constants');
+const { AtbashCipherTransphormStream,
+        CaesarCipherTransphormStream,
+        Rot8CipherTransphormStream,
+        CustomWritableStream,
+        CustomReadableStream,
+        customStdin } = require('./stream');
 
 try {
   const initialArgs = process.argv.slice(2);
@@ -15,18 +15,11 @@ try {
   validateOptions(initialArgs);
   validateConfig(config);
 
-  // const readableStream = input ? fs.createReadStream(`${input}`) : customStdin();
   const readableStream = input ? new CustomReadableStream(input) : customStdin();
   const writableStream = output ? new CustomWritableStream(output) : stdout;
   const transformStream = new CipherTransphorm(config);
-  // const handleReadError = () => {
-  //   readableStream.destroy();
-  //   stderr.write(ERROR_MESSAGE.INPUT);
-  //   exit(1);
-  // };
 
   readableStream
-    // .on('error', handleReadError)
     .pipe(transformStream)
     .pipe(writableStream);
 }
