+++ stream/index.js
@@ -1,0 +1,15 @@
+const AtbashCipherTransphormStream = require('./atbashCipherTransphormStream');
+const CaesarCipherTransphormStream = require('./caesarCipherTransphormStream');
+const Rot8CipherTransphormStream = require('./rot8CipherTransphormStream');
+const CustomWritableStream = require('./customWritableStream');
+const CustomReadableStream = require('./customReadableStream');
+const customStdin = require('./customStdin');
+
+module.exports = {
+  AtbashCipherTransphormStream,
+  CaesarCipherTransphormStream,
+  Rot8CipherTransphormStream,
+  CustomWritableStream,
+  CustomReadableStream,
+  customStdin
+}
